# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/transactions": {
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      },
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "checkout",
                "credit"
              ],
              "type": "string"
            }
          },
          {
            "name": "accountIdList",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "categoryIdList",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryTransactionOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      }
    },
    "/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      },
      "patch": {
        "operationId": "TransactionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      },
      "delete": {
        "operationId": "TransactionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTransactionOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Transactions"
        ]
      }
    },
    "/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCategoryOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
    "/categories/{id}/appearance": {
      "patch": {
        "operationId": "CategoriesController_updateAppearance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryAppearanceInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
    "/groups": {
      "post": {
        "operationId": "GroupsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ]
      },
      "get": {
        "operationId": "GroupsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupListOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
    "/groups/{id}": {
      "get": {
        "operationId": "GroupsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ]
      },
      "delete": {
        "operationId": "GroupsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteGroupInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteGroupOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Groups"
        ]
      }
    },
    "/bankaccounts": {
      "post": {
        "operationId": "BankaccountsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBankaccountInputDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankaccountOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Bankaccounts"
        ]
      },
      "get": {
        "operationId": "BankaccountsController_findAll",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankaccountListOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Bankaccounts"
        ]
      }
    },
    "/bankaccounts/{id}": {
      "delete": {
        "operationId": "BankaccountsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteBankaccountOutputDto"
                }
              }
            }
          }
        },
        "tags": [
          "Bankaccounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Fina API",
    "description": "API for the Fina financial collaboration app",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token here",
        "in": "header"
      }
    },
    "schemas": {
      "CreateTransactionInputDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number",
            "nullable": true
          },
          "date": {
            "nullable": true
          },
          "installmentTotal": {
            "type": "integer",
            "nullable": true
          },
          "installmentCurrent": {
            "type": "string",
            "nullable": true
          },
          "observation": {
            "type": "string",
            "nullable": true
          },
          "bankaccountId": {
            "type": "string",
            "pattern": "^\\d+$"
          },
          "categoryId": {
            "type": "string",
            "pattern": "^\\d+$",
            "nullable": true
          },
          "groupId": {
            "type": "string",
            "pattern": "^\\d+$"
          },
          "importId": {
            "type": "string",
            "pattern": "^\\d+$",
            "nullable": true
          },
          "billMonth": {
            "type": "string",
            "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])$",
            "nullable": true
          }
        },
        "required": [
          "description",
          "value",
          "date",
          "bankaccountId",
          "groupId",
          "billMonth"
        ]
      },
      "TransactionOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "installmentTotal": {
            "type": "number",
            "nullable": true
          },
          "installmentCurrent": {
            "type": "string",
            "nullable": true
          },
          "creditDueDate": {
            "type": "string",
            "nullable": true
          },
          "observation": {
            "type": "string",
            "nullable": true
          },
          "toBeConsideredAt": {
            "type": "string",
            "nullable": true
          },
          "calculatedDate": {
            "type": "string",
            "nullable": true
          },
          "billPayment": {
            "type": "object",
            "properties": {
              "creditAccountId": {
                "type": "string"
              },
              "billMonth": {
                "type": "string",
                "pattern": "^\\d{4}-(?:0[1-9]|1[0-2])$"
              }
            },
            "required": [
              "creditAccountId",
              "billMonth"
            ],
            "nullable": true
          },
          "bankaccount": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "type": {
                "type": "string"
              },
              "dueDate": {
                "type": "integer",
                "minimum": 1,
                "exclusiveMinimum": false,
                "maximum": 31,
                "exclusiveMaximum": false,
                "nullable": true
              }
            },
            "required": [
              "id",
              "name",
              "type",
              "dueDate"
            ],
            "nullable": true
          },
          "category": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "icon": {
                "type": "string",
                "enum": [
                  "tag",
                  "shopping-cart",
                  "utensils",
                  "car",
                  "house",
                  "heart-pulse",
                  "plane",
                  "gamepad-2",
                  "briefcase",
                  "book-open",
                  "coffee",
                  "dumbbell",
                  "gift",
                  "paw-print",
                  "shirt",
                  "smartphone",
                  "music",
                  "baby",
                  "wrench",
                  "wallet-cards",
                  "bike",
                  "camera",
                  "graduation-cap",
                  "sparkles"
                ]
              },
              "color": {
                "type": "string",
                "enum": [
                  "yellow",
                  "lime",
                  "sky",
                  "violet",
                  "danger",
                  "amber",
                  "orange",
                  "rose",
                  "pink",
                  "fuchsia",
                  "indigo",
                  "teal"
                ]
              }
            },
            "required": [
              "id",
              "name",
              "icon",
              "color"
            ],
            "nullable": true
          },
          "group": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ]
          },
          "import": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "fileName": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "fileName"
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "createdAt",
          "description",
          "value",
          "date",
          "installmentTotal",
          "installmentCurrent",
          "creditDueDate",
          "observation",
          "toBeConsideredAt",
          "calculatedDate",
          "billPayment",
          "bankaccount",
          "category",
          "group",
          "import"
        ]
      },
...
      "UpdateTransactionInputDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number",
            "nullable": true
          },
          "date": {
            "nullable": true
          },
          "installmentTotal": {
            "type": "integer",
            "nullable": true
          },
          "installmentCurrent": {
            "type": "string",
            "nullable": true
          },
          "creditDueDate": {
            "nullable": true
          },
          "observation": {
            "type": "string",
            "nullable": true
          },
          "toBeConsideredAt": {
            "nullable": true
          },
          "calculatedDate": {
            "nullable": true
          },
          "bankaccountId": {
            "type": "string",
            "pattern": "^\\d+$",
            "nullable": true
          },
          "categoryId": {
            "type": "string",
            "pattern": "^\\d+$",
            "nullable": true
          }
        }
      },
      "DeleteTransactionOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "CreateCategoryInputDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "icon": {
            "type": "string",
            "enum": [
              "tag",
              "shopping-cart",
              "utensils",
              "car",
              "house",
              "heart-pulse",
              "plane",
              "gamepad-2",
              "briefcase",
              "book-open",
              "coffee",
              "dumbbell",
              "gift",
              "paw-print",
              "shirt",
              "smartphone",
              "music",
              "baby",
              "wrench",
              "wallet-cards",
              "bike",
              "camera",
              "graduation-cap",
              "sparkles"
            ],
            "default": "tag"
          },
          "color": {
            "type": "string",
            "enum": [
              "yellow",
              "lime",
              "sky",
              "violet",
              "danger",
              "amber",
              "orange",
              "rose",
              "pink",
              "fuchsia",
              "indigo",
              "teal"
            ],
            "default": "yellow"
          },
          "groupId": {
            "type": "string",
            "pattern": "^\\d+$"
          }
        },
        "required": [
          "name",
          "groupId"
        ]
      },
      "CategoryOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string",
            "enum": [
              "tag",
              "shopping-cart",
              "utensils",
              "car",
              "house",
              "heart-pulse",
              "plane",
              "gamepad-2",
              "briefcase",
              "book-open",
              "coffee",
              "dumbbell",
              "gift",
              "paw-print",
              "shirt",
              "smartphone",
              "music",
              "baby",
              "wrench",
              "wallet-cards",
              "bike",
              "camera",
              "graduation-cap",
              "sparkles"
            ]
          },
          "color": {
            "type": "string",
            "enum": [
              "yellow",
              "lime",
              "sky",
              "violet",
              "danger",
              "amber",
              "orange",
              "rose",
              "pink",
              "fuchsia",
              "indigo",
              "teal"
            ]
          },
          "groupId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "name",
          "icon",
          "color",
          "groupId"
        ]
      },
...
      "DeleteCategoryOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "CreateGroupInputDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          }
        },
        "required": [
          "name"
        ]
      },
...
      "DeleteGroupInputDto": {
        "type": "object",
        "properties": {
          "confirmName": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "confirmName"
        ]
      },
...
    }
  },
  "security": [
    {
      "access-token": []
    }
  ]
}